Split external and pseudo units' implementations into files.
