+++ flang/runtime/io-stmt.cpp
@@ -227,12 +227,11 @@ ConnectionState &ExternalIoStatementBase::GetConnectionState() { return unit_; }
 int ExternalIoStatementBase::EndIoStatement() {
   CompleteOperation();
   auto result{IoStatementBase::EndIoStatement()};
-#if defined(RT_USE_PSEUDO_FILE_UNIT)
+#if !defined(RT_USE_PSEUDO_FILE_UNIT)
+  unit_.EndIoStatement(); // annihilates *this in unit_.u_
+#else
   // Fetch the unit pointer before *this disappears.
   ExternalFileUnit *unitPtr{&unit_};
-#endif
-  unit_.EndIoStatement(); // annihilates *this in unit_.u_
-#if defined(RT_USE_PSEUDO_FILE_UNIT)
   // The pseudo file units are dynamically allocated
   // and are not tracked in the unit map.
   // They have to be destructed and deallocated here.
